2023 data run

+++ Scripts/yolov8m_seg_detection_workflow.py
@@ -19,8 +19,7 @@ def create_output_folders(base_dir, output_dir, draw, conf_threshold):
         output_dir = base_dir
     reproc_dir = os.path.join(output_dir, "REPROC")
     target_base = reproc_dir if os.path.isdir(reproc_dir) else output_dir
-    timestamp = datetime.datetime.now().strftime("%Y_%m_%d_%H_%M")
-    timestamp_conf = f"{timestamp}_CONF_{str(int(conf_threshold*100))}"
+    timestamp_conf = f"{datetime.datetime.now():%Y%m%d_%H%M}_C{int(conf_threshold*100)}"
     run_dir = os.path.join(target_base, timestamp_conf)
     tile_dir = os.path.join(run_dir, "TILES")
     os.makedirs(tile_dir, exist_ok=True)
@@ -71,7 +70,7 @@ def compute_overlap_metrics(boxA, boxB):
 
     return iou, containment, areaA, areaB
 
-def merge_detections(detections, iou_thresh=0.6, containment_thresh=0.85):
+def merge_detections(detections, iou_thresh=0.6, containment_thresh=0.9):
     detections = sorted(detections, key=lambda d: (
         (d['bbox'][2] - d['bbox'][0]) * (d['bbox'][3] - d['bbox'][1])
     ), reverse=True)
@@ -146,7 +145,7 @@ def get_lat_lon(image_path):
 
 def process_single_image(img_path, model_path, conf_threshold, draw, run_dir, detect_dir):
     img_name = os.path.basename(img_path)
-    tile_dir = os.path.join(run_dir, "TILES", f"{Path(img_name).stem}_{uuid.uuid4().hex[:8]}")
+    tile_dir = os.path.join(run_dir, "TILES", uuid.uuid4().hex[:8])
     os.makedirs(tile_dir, exist_ok=True)
 
     model = YOLO(model_path)
@@ -160,7 +159,7 @@ def process_single_image(img_path, model_path, conf_threshold, draw, run_dir, de
     detections = []
 
     for idx, ((x_off, y_off), tile) in enumerate(tiles):
-        tile_path = os.path.join(tile_dir, f"{Path(img_name).stem}_tile_{idx}.jpg")
+        tile_path = os.path.join(tile_dir, f"t{idx}.jpg")
         cv2.imwrite(tile_path, tile)
         results = model(tile_path, verbose=False)[0]
 
@@ -196,7 +195,8 @@ def process_single_image(img_path, model_path, conf_threshold, draw, run_dir, de
         row = [det['image'], latitude, longitude, det['confidence']] + [f"({x},{y})" for (x, y) in det['corners']]
         detections_csv.append(row)
 
-    shutil.rmtree(tile_dir)
+    if os.path.exists(tile_dir):
+        shutil.rmtree(tile_dir)
 
     return detections_csv
 
@@ -248,9 +248,9 @@ def process_images(img_dir, model_dir, conf_threshold, draw=True, output_dir=Non
 
 if __name__ == "__main__":
     image_dir = "Sample_Images/"
-    model_path = "Models/seal-segmentation-v2-1/weights/best.pt"
+    model_path = "Models/seal-box-v3-1/weights/best.pt"
     output_dir = None
-    conf_threshold = 0.65
+    conf_threshold = 0.8
     draw = True
 
     csv_file = process_images(image_dir, model_path, conf_threshold, draw, output_dir)
